Migrate Highlights component to plain React
Migrates the Highlights main component from styled-components to plain React
while maintaining backward compatibility.

Changes:
- Highlights.new.tsx: Plain React implementation without styled-components
  - Removed dependency on styled-components/macro
  - Component logic unchanged, just removing the styled() wrapper

- Highlights.legacy.tsx: Styled-components wrapper for backward compatibility
  - Maintains the print styles that use component selectors
  - Targets HighlightWrapper from SectionHighlights using ${HighlightWrapper}
  - Required until SectionHighlights is migrated to plain CSS

- Highlights.css: Placeholder CSS file for future plain CSS print styles
  - Currently not used because HighlightWrapper is still a styled component
  - Will be activated when SectionHighlights is migrated

- Highlights.tsx: Re-export shim maintaining original import path

Key Notes:
The legacy wrapper is still required because it uses styled-component
selectors to target HighlightWrapper from SectionHighlights.tsx. Once
SectionHighlights is migrated, the print styles can be converted to plain
CSS using a .highlight-wrapper class selector.

Related: CORE-2282
