+++ src/app/content/highlights/components/Highlights.css
@@ -1,0 +1,13 @@
+/**
+ * Highlights Component Styles
+ *
+ * This file contains print-specific styles for the main Highlights component.
+ * It handles margin adjustments for HighlightWrapper elements when printing.
+ */
+
+/* === Print Styles === */
+@media print {
+  .highlights-component .highlight-wrapper {
+    margin: 0;
+  }
+}
